// src/lib/api/api.ts
export interface components {
    schemas: {
// ...
        DCustomer: {
            /** Format: int64 */
            id?: number;
            isActive?: boolean;
            /** Format: date-time */
            createdDate?: string;
            /** Format: date-time */
            updatedDate?: string;
            firstName?: string;
            lastName?: string;
            email?: string;
            phoneNumber?: string;
            /** Format: date */
            birthDate?: string;
            /** @enum {string} */
            idDocumentType?: "PASSPORT" | "DNI" | "DRIVER_LICENSE" | "RUC" | "CE";
            idDocumentNumber?: string;
            address?: string;
            nationality?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}